fix: changed result value to BigInteger

+++ job-service/src/main/java/com/fl1mt/jobservice/domain/Job.java
@@ -7,6 +7,7 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Entity
@@ -27,7 +28,7 @@ public class Job {
     @Column(name = "job_status")
     private JobStatus status;
     @Column(name = "result", nullable = false)
-    private Long result;
+    private BigInteger result;
     @Column(name = "created_at", updatable = false)
     @CreatedDate
     private LocalDateTime createdAt;
